test(marshal,exo): stop exporting from test-*.js files
