+++ packages/marshal/test/test-marshal-stringify.js
@@ -3,7 +3,7 @@ import { test } from './prepare-test-env-ava.js';
 // eslint-disable-next-line import/order
 import { Far } from '@endo/pass-style';
 import { stringify, parse } from '../src/marshal-stringify.js';
-import { roundTripPairs } from './test-marshal-capdata.js';
+import { roundTripPairs } from './marshal-test-data.js';
 
 const { isFrozen } = Object;
